Fixing some problems to make it build on my system
The gradle build file compiled the image slider in a funny way (from local folder maybe?)
so I just changed it to compile according to the recommendation on the image slider github
page. Also, not so far from gucci's 'image 5 bug', my [newer?] version of image slider
demanded a string rather than a bitmap for a function. I just put in the string, result
as a placeholder so that the app builds and I can start.